Add battery voltage bar to flight controls tab:
Adds a new bar to the right of the M1-M4 bars showing the battery voltage
as reported by the pm.vbat log variable, scaled between the hardcoded
variables BATTERY_VOLTAGE_MIN and BATTERY_VOLTAGE_MAX. I got the value for
the latter from the firmware source, and the former was determined via
experimentation. Future work is to retrieve these values from the copter
itself somehow.

The bar color is green when charging, red when in the low-power state,
and blue otherwise. Unfortunately, styling the progress bar this way
means we lose the pretty native styles, but I believe it's worth it.

This uncovered an odd behavior: when fully charged, the copter reports
the pm.state variable as "battery", rather than "charging" or "charged",
as would be expected (see possible values in hal/interface/pm.h in the
firmware source).