Change battery bar color based on state (charging, draining, low-power):
The bar color is green when charging, red when in the low-power state,
and blue otherwise. Unfortunately, styling the progress bar this way
means we lose the pretty native styles, but I believe it's worth it.

This uncovered an odd behavior: when fully charged, the copter reports
the pm.state variable as battery, rather than charging or charged, as
would be expected (see possible values in hal/interface/pm.h in the
firmware source).