Fixed blank UI issue (at least on Macs)
Often the UI would not refresh/repaint and the user had to force a
repaint by either clicking on every component or triggering a repaint
(eg: maximizing the window). Now, a QTimer with timeout=0 (which means
it will be called whenever the event loop is available/done processing)
triggers a repaint and UI issues are gone! :)