Better way to force repaint on special widgets
Toggling the visibility of the special widgets (such as the param list,
plotter, console output...) twice (hide + show) solved the "blank UI"
issue. However, I noticed that the selected widget in the active tab
would lose focus every time the timer_repaint toggled the visibility of
such special widgets. This was especially annoying in tabs that contain
user input widgets (for example, the Plotter tab allows the user to
manually select the range for the plot axes), which would lose focus
(every 500ms) while the user was typing.
The workaround this time is to resize those special widgets by making
them 1px smaller and resizing back to the original size. This triggers
the internal repaint that I haven't been able to call explicitly, but
at least it works! :)