FlightTab: Add Command Based Flight Control
If connected and a positioning deck (bcFlow, bcFlow2, bcLightHouse4,
bcDWM1000) is found we enable a simple interface to send high level
commands to the Crazyflie.

This is to make the "first encounter" with the Crazyflie and the client
more enjoyable, and to enable a user to fly around with the drone
without any real experience.