main: warn once about USB permissions on scan
When the user press the scan button and we get a Permissions Denied
error from the USB core then we know that the Crazyradio dongle is
present, but we cannot access it.

We will warn, once, about this with a message box that gives a link
to instructions on how to install the udev files.

Closes #444