Implement x axis manipulation in plotting tab
The last N seconds can now be plotted.
A range can now be set in seconds (low, high limits). The range is automatically adapted to match the latest view in samples / seconds mode. This makes the range mode useful for zooming in on certain parts of the view.

Changed x axis behavior such that the range configuration (samples, seconds, range) dictates x limits of the plot even if there is a (partial) lack of data for the range. Rather than stretching the plot as it fills up, now the entire plot limits are shown immediately. This immediate feedback helps with choosing a desirable x axis setting.